+++ tests/unit/test_nats_adapter.py
@@ -56,10 +56,12 @@ class TestNatsKVAdapter:
     async def test_connect(self, mock_nats_kv_adapter):
         mock_nats_kv_adapter._js.key_value = AsyncMock(side_effect=BucketNotFoundError)
 
-        result = await mock_nats_kv_adapter.init([Bucket.subscriptions])
+        result = await mock_nats_kv_adapter.init(
+            [Bucket.subscriptions], user="test_user", password="test_password"
+        )
 
         mock_nats_kv_adapter._nats.connect.assert_called_once_with(
-            ["nats://localhost:4222"]
+            ["nats://localhost:4222"], user="test_user", password="test_password"
         )
         mock_nats_kv_adapter._js.create_key_value.assert_called_once_with(
             bucket=Bucket.subscriptions
@@ -133,10 +135,12 @@ class TestNatsKVAdapter:
 @pytest.mark.anyio
 class TestNatsMQAdapter:
     async def test_connect(self, mock_nats_mq_adapter):
-        result = await mock_nats_mq_adapter.connect()
+        result = await mock_nats_mq_adapter.connect(
+            user="test_user", password="test_password"
+        )
 
         mock_nats_mq_adapter._nats.connect.assert_called_once_with(
-            ["nats://localhost:4222"]
+            ["nats://localhost:4222"], user="test_user", password="test_password"
         )
         assert result is None